requirements.txt
requests>=2.28.0       # for sending HTTP to LM Studio
pytest>=7.0.0          # for unit testing
flake8>=6.0.0          # for linting and CI checks
PyQt6>=6.4.0           # for GUI dashboard
python-docx>=0.8.11    # for processing .docx files
llama-cpp-python>=0.3.9 # for local LLM inference with Phi-2
